Combining the CLI and GUI version into one main and giving the user the choice which to use
